Oh boy this is a big one. Changes: -Compiler changed from v120ctp to v140 -Reworked PAD model format. Model converter is much smaller as well. -Reworked model resource handling. Everything is loaded piece-wise and on demand, rather than the entire model. -Reworked model loader. Now only specific sections may be loaded into memory. In addition, the format is now laid out in structs for easy update and IO. -Reworked animation system. No, like really reworked it: --Animation sampling removed from renderer and placed into physics step. --Transform injection removed from the millions of places and consolidated into animator --Animations may now sample from multiple animation sets. Magnificus! --Animations are exported into a set of matrices once per frame, as opposed to the gazillions of recalculations previously. -Reworked skinning system. Meshes will attempt to reuse the same buffer, and will stream the data at the end of the frame to fix AMD latency issues. -Reworked skeletal system. No longer relies on Transform, using new LiteTransform and monotonic parenting arrays. -Morpher mesh streaming greatly optimized, only using a single new buffer instead of leaking buffers all over the damn place. -glDrawing reworked to accept floats when drawing by pixel coordinate. No sense in casting from a float to an int to a float when we round the first float anyways. -glSkinnedMeshBone aka glBone is now deprecated and should never be used ever again. -CBufferIO now transfers ownership of buffers properly, and will deallocate in a way that makes sense on destruction -CSegmentedFile is slightly optimized to not be a turd when seeking to specific sections, and will return a CBufferIO instead of a std::string (whose idea was that?!) -CBinaryFile was including "boost/mutex.hpp" which was a file that included <mutex>. What?!? New: -Added Core::LiteTransform class for your light transform needs! Use for quicky parenting calculations without touching a damn matrix. -BoneMapper for mapping one Skeleton to another! -ModelLoader for loading specific model data. -DukScript, an in-progress scripting integration for readable and maintainable scripting. -StringUtils has new (and very thoroughly tested) string comparison functions useful for DNA processing...or creating BoneMapper instances. That too. -XrTransform, the popular new cousin of XTransform and XvTransform, with its 64 bytes of fully functional badassery, fully capable of organizing messy transformation code Bugs: -AFTER is currently not compiling. New animation and skeletal system and such. Simple changes. -Facial morphs don't load properly. This is likely due to the way morphs are not handled at all ever

+++ M01_AFTER/oneGame/source/after-editor/scenes/gmsceneLipsyncEditor.cpp
@@ -224,7 +224,7 @@ void gmsceneLipsyncEditor::LoadScene ( void )
 		tattoo.projection_scale = Vector3d(1,1,1) * 0.83f;
 
 		rstats->tattooList.push_back( tattoo );*/
-		rstats->iTattooCount = rstats->tattooList.size();
+		rstats->iTattooCount = (uint16_t) rstats->tattooList.size();
 	}
 	loadscreen->StepScreen();
